rollup improvements and bugfixes
