+++ src/ktapi.js
@@ -1,89 +1,13 @@
-var baseUrl = "https://kt.125m125.de/api/v2.0/",
-    paramsToQuery = function (params) {
-        if (!params) {
-            return params;
-        }
-        var result = "",
-            sorted_keys = Object.keys(params).sort();
-        sorted_keys.forEach(function (entry) {
-            result += entry + "=" + params[entry] + "&";
-        });
-        if (result.length) {
-            return result.slice(0, -1);
-        }
-        return result;
-    },
-    performRequest = function (method, type, suburl, params, headers, callback) {
-        var url, request;
-        if (method === "GET") {
-            url = baseUrl + suburl + "?" + paramsToQuery(params);
-            params = null;
-        } else {
-            url = baseUrl + suburl;
-            if (!headers["Content-Type"]) {
-                headers["Content-Type"] = "application/x-www-form-urlencoded";
-            }
-        }
-        request = d3[type](url);
-        Object.keys(headers).forEach(function (key) {
-            request.header(key, headers[key]);
-        });
-        request.on("error", function (error) {
-            var errorTarget = error.target;
-            try {
-                return callback(JSON.parse(errorTarget.response));
-            } catch (e) { }
-            callback(errorTarget);
-        }).on("load", function (data) {
-            callback(false, data);
-        });
-        request.send(method, paramsToQuery(params));
-    };
-
-function Request(method, type, suburl, params, headers, user, authenticator) {
-    this.execute = function (callback) {
-        if (!params) {
-            params = {};
-        }
-        if (!headers) {
-            headers = {};
-        }
-        suburl = suburl.replace(/\{[a-zA-Z0-9]+?\}/g, function (key) {
-            if (key === "{user}") {
-                return user.uid;
-            }
-            var key2 = key.substring(1, key.length - 1),
-                result;
-            if (params.hasOwnProperty(key2) && params[key2] !== undefined) {
-                result = params[key2];
-                delete params[key2];
-                return result;
-            }
-            return key;
-        });
-        if (authenticator) {
-            authenticator.authenticate(method, type, suburl, params, headers, user);
-        }
-        performRequest(method, type, suburl, params, headers, callback);
-    };
-}
+import { Request } from './requests/Request.js';
 
 var KtApi = function (uid, tid, tkn, authenticator) {
     var user = {
         "uid": uid,
         "tid": tid,
         "tkn": tkn
-    },
-        permissions;
-
-    this.getPermissions = function (callback) {
-        var start = new Date();
-        performRequest("GET", "json", "permissions", {
-            uid: user.uid,
-            tid: user.tid
-        }, user.tkn, false, callback);
     };
-    this.getRequest = function (method, type, suburl, params, headers, auth) {
+
+   this.getRequest = function (method, type, suburl, params, headers, auth) {
         var requestAuthenticator;
         if (auth) {
             requestAuthenticator = authenticator;
@@ -94,6 +18,11 @@ var KtApi = function (uid, tid, tkn, authenticator) {
         return this.getRequest(method, type, suburl, params, headers, auth).execute(callback);
     };
 }
+
+KtApi.prototype.getPermissions = function (callback) {
+    this.performRequest("GET", "json", "permissions/{user}", null, null, true, callback);
+};
+
 KtApi.prototype.getItems = function (callback) {
     this.performRequest("GET", "tsv", "users/{user}/items", null, null, true, callback);
 };
@@ -154,7 +83,7 @@ KtApi.prototype.pusherAuthenticate = function (socket, channel, callback) {
 };
 
 KtApi.prototype.getPayouts = function (callback) {
-    performRequest("GET", "tsv", "users/{user}/payouts", null, null, true, callback);
+    this.performRequest("GET", "tsv", "users/{user}/payouts", null, null, true, callback);
 };
 
 KtApi.prototype.createPayout = function (type, item, count, callback) {
